+++ examples/booktest/cpp/native/queries.cpp
@@ -413,7 +413,7 @@ void Queries::RemoveBookTag(std::uint64_t book_id, const std::string& tag) const
             "UPDATE books\n"
             "SET tags = UNWRAP(Yson::SerializeJson(Json::From(ListFilter(\n"
             "    Yson::ConvertToStringList(tags),\n"
-            "    ($item) -> ($item != $tag)\n"
+            "    ($item) -> ($item NOT IN AsList($tag))\n"
             "))))\n"
             "WHERE book_id = $book_id;",
             sqlc_tx,