+++ examples/booktest/go/database/sql/queries.sql.go
@@ -227,7 +227,7 @@ func (q *Queries) RemoveBookTag(ctx context.Context, arg RemoveBookTagParams) er
 		"UPDATE books\n"+
 		"SET tags = UNWRAP(Yson::SerializeJson(Json::From(ListFilter(\n"+
 		"    Yson::ConvertToStringList(tags),\n"+
-		"    ($item) -> ($item != $tag)\n"+
+		"    ($item) -> ($item NOT IN AsList($tag))\n"+
 		"))))\n"+
 		"WHERE book_id = $book_id;",
 		sql.Named("book_id", arg.BookID),